requirements.txt
basicsr
realesrgan
pillow
numpy
torch